replaced state.State with state.MerkleState
