+++ vms/platformvm/state/state_helpers.go
