+++ vms/platformvm/state/state_helpers_test.go
