+++ vms/platformvm/txs/executor/helpers_test.go
@@ -225,7 +225,7 @@ func defaultState(
 ) state.State {
 	genesisBytes := buildGenesisTest(ctx)
 	execCfg, _ := config.GetExecutionConfig(nil)
-	state, err := state.NewMerkleState(
+	state, err := state.New(
 		db,
 		genesisBytes,
 		prometheus.NewRegistry(),
