+++ vms/platformvm/vm_regression_test.go
@@ -644,7 +644,7 @@ func TestRejectedStateRegressionInvalidValidatorTimestamp(t *testing.T) {
 	// Force a reload of the state from the database.
 	vm.Config.Validators = validators.NewManager()
 	execCfg, _ := config.GetExecutionConfig(nil)
-	newState, err := state.NewMerkleState(
+	newState, err := state.New(
 		vm.db,
 		nil,
 		prometheus.NewRegistry(),
@@ -951,7 +951,7 @@ func TestRejectedStateRegressionInvalidValidatorReward(t *testing.T) {
 	// Force a reload of the state from the database.
 	vm.Config.Validators = validators.NewManager()
 	execCfg, _ := config.GetExecutionConfig(nil)
-	newState, err := state.NewMerkleState(
+	newState, err := state.New(
 		vm.db,
 		nil,
 		prometheus.NewRegistry(),
